back/requirements.txt
asgiref==3.11.0
bleach==6.1.0
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3
django-ratelimit>=4.1.0
Pillow>=10.2.0
psycopg2-binary==2.9.9
sqlparse==0.5.5
tzdata==2025.3
google-generativeai>=0.5.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-core>=0.1.0
celery>=5.3.0
redis>=5.0.0
django-redis>=5.4.0
numpy>=1.26.0
filetype>=1.2.0  # File type detection (pure Python, cross-platform)
drf-spectacular>=0.27.0  # OpenAPI/Swagger documentation
channels>=4.0.0  # WebSocket support
channels-redis>=4.1.0  # Redis channel layer
daphne>=4.0.0  # ASGI server
cryptography>=46.0.0  # Encryption for webhooks
sentry-sdk>=1.40.0  # Error tracking and performance monitoring
django-health-check>=3.17.0  # Health check endpoint for monitoring
